backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JSONRPC==2.2.2
gunicorn==21.2.0
langchain==0.0.314
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.5.1
